// src/models/content-types.ts
export enum ContentTypes {
    BINARY = 'application/octet-stream',
    BZIP = 'application/x-bzip',
    BZIP2 = 'application/x-bzip2',
    CSV = 'text/csv',
    GZIP = 'application/gzip',
    JSON = 'application/json',
    PDF = 'application/pdf',
    RAR = 'application/vnd.rar',
    TAR = 'application/x-tar',
    TEXT = 'text/plain',
    XLSX = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
    XML = 'application/xml',
    ZIP = 'application/zip',
    _7ZIP = 'application/x-7z-compressed'
}